refactor: relocate createAlert method
