--- MovieQuiz/Services/Gamerecord.swift
@@ -1,32 +1,0 @@
-//
-//  Gamerecord.swift
-//  MovieQuiz
-//
-//  Created by Yulianna on 26.01.2024.
-//
-
-import Foundation
-struct GameRecord: Codable {
-    private enum Keys: CodingKey {
-        case correct, total, date
-    }
-    let correct: Int
-    let total: Int
-    let date: Date
-    init(correct: Int, total: Int, date: Date) {
-        self.correct = correct
-        self.total = total
-        self.date = date
-    }
-    
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: Keys.self)
-        self.correct = try container.decode(Int.self, forKey: .correct)
-        self.total = try container.decode(Int.self, forKey: .total)
-        self.date = try container.decode(Date.self, forKey: .date)
-    }
-    func isBetterThan(_ another: GameRecord) -> Bool {
-        correct < another.correct
-    }
-}
-
